requirements.txt
adrf
africastalking
aiohappyeyeballs
aiohttp
aiosignal
amqp
anyio
asgiref
async-property
async-timeout
attrs
autobahn
Automat
azure-core
azure-identity
bandit
billiard
boto
boto3
cachetools
celery
certifi
cffi
channels
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
constantly
contextlib2
coverage
cron-descriptor
cryptography
daphne
defusedxml
Deprecated
diff-match-patch
Django
django-braces
django-celery-beat
django-celery-results
django-ckeditor
django-ckeditor-5
django-cors-headers
django-filter
django-import-export
django-jazzmin
django-js-asset
django-redis
django-storages
django-timezone-field
djangorestframework
djangorestframework-simplejwt
drf-spectacular
drf-spectacular-sidecar
drf-yasg
et-xmlfile
exceptiongroup
frozenlist
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
gunicorn
h11
h2
hpack
httpcore
httplib2
httptools
httpx
hyperframe
hyperlink
idna
incremental
inflection
jmespath
jsonschema
jsonschema-specifications
kombu
lxml
Markdown
markdown-it-py
mdurl
microsoft-kiota-abstractions
microsoft-kiota-authentication-azure
microsoft-kiota-http
microsoft-kiota-serialization-form
microsoft-kiota-serialization-json
microsoft-kiota-serialization-multipart
microsoft-kiota-serialization-text
msal
msal-extensions
msgraph-core
msgraph-sdk
multidict
numpy
oauth2client
oauthlib
openpyxl
opentelemetry-api
opentelemetry-sdk
opentelemetry-semantic-conventions
packaging
pandas
pbr
pendulum
pillow
portalocker
priority
prompt_toolkit
proto-plus
protobuf
# psycopg2
psycopg2-binary
pyasn1
pyasn1_modules
pycountry
pycparser
Pygments
PyJWT
pyOpenSSL
pyparsing
python-crontab
python-dateutil
python-docx
python-dotenv
python-pptx
pytz
PyYAML
redis
referencing
requests
requests-oauthlib
rich
rpds-py
rsa
s3transfer
schema
sentry-sdk
service-identity
setuptools
six
sniffio
sqlparse
std-uritemplate
stevedore
time-machine
Twisted
txaio
typing_extensions
tzdata
uritemplate
uvicorn[standard]
uvloop
vine
watchfiles
wcwidth
websockets
whitenoise
wrapt
XlsxWriter
yarl
zipp
zope.interface
qrcode
supervisor
